Use `usize` instead of `isize`.
`memchr` should gurantee that a non-null pointer returned is strictly
greater than or equal to the start pointer, so this should never overflow.

+++ src/lib.rs
@@ -40,7 +40,7 @@ pub fn memchr(needle: u8, haystack: &[u8]) -> Option<usize> {
     if p.is_null() {
         None
     } else {
-        Some((p as isize - (haystack.as_ptr() as isize)) as usize)
+        Some(p as usize - (haystack.as_ptr() as usize))
     }
 }
